requirements.txt
mechanize
bs4
html2text
Wand
Pillow
discord.py
requests
numpy
lxml
